# Conflicts:
#	src/main/java/at/hannibal2/skyhanni/api/enoughupdates/EnoughUpdatesManager.kt
#	src/main/java/at/hannibal2/skyhanni/api/pet/CurrentPetApi.kt
#	src/main/java/at/hannibal2/skyhanni/api/storage/StorageApi.kt
#	src/main/java/at/hannibal2/skyhanni/data/IslandType.kt
#	src/main/java/at/hannibal2/skyhanni/events/entity/EntityClickEvent.kt
#	src/main/java/at/hannibal2/skyhanni/features/inventory/storage/ItemTagManager.kt
#	src/main/java/at/hannibal2/skyhanni/utils/NeuNPC.kt
#	src/main/java/at/hannibal2/skyhanni/utils/PrimitiveIngredient.kt
#	src/main/java/de/hype/bingonet/shared/constants/Islands.kt
#	src/main/java/de/hype/bingonet/shared/constants/Skills.kt
#	src/main/java/de/hype/bingonet/sharedcompilation/sbenums/BNNEUItem.kt

+++ src/main/java/at/hannibal2/skyhanni/utils/renderables/RenderableUtils.kt
@@ -25,6 +25,7 @@ import java.awt.Color
 import kotlin.math.ceil
 import kotlin.math.min
 import kotlin.reflect.KMutableProperty0
+
 //#if MC > 1.21
 //$$ import net.minecraft.text.Text
 //#endif
@@ -425,7 +426,7 @@ internal object RenderableUtils {
         data: List<DisplayTableEntry>,
         itemScale: Double = NeuItems.ITEM_FONT_SIZE,
     ): MutableList<List<Renderable>> {
-        val sorted = data.sortedByDescending { it.sort }
+        val sorted = data.sortedByDescending { it.sort.toDouble() }
         val outerList = mutableListOf<List<Renderable>>()
         for (entry in sorted) {
             val item = entry.item.getItemStackOrNull()?.let {
